chatbot/requirements.txt
streamlit>=1.28.0
huggingface-hub>=0.20.0
scikit-learn>=1.0.0
python-dotenv>=1.0.0
numpy>=1.0.0
